# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "mitmproxy Grid",
        "description": "\n## mitmproxy Grid API\n\nManage a fleet of containerised mitmproxy instances, each running on its own\nport with independent interception rules and CA certificates.\n\n### Instance lifecycle\nEvery instance has a configurable **TTL** (default `1800s` via the\n`INSTANCE_TTL` environment variable).  Expired instances are automatically\ndestroyed.  Use **POST /instances/{{id}}/renew** to extend an instance.\n\n### Rule model (v2)\nRules separate **match criteria** from **actions**.\nA single rule can modify both the request and the response:\n\n```json\n{{\n  \"match\": {{ \"urlPattern\": \".*api.*\", \"method\": \"GET\" }},\n  \"action\": {{\n    \"modifyRequest\":  {{ \"headers\": {{ \"set\": {{ \"X-Debug\": \"1\" }} }} }},\n    \"modifyResponse\": {{ \"body\": \"{{\\\"mocked\\\": true}}\" }}\n  }}\n}}\n```\n\n### Quick start\n1. `POST /instances` \u2192 spin up a proxy\n2. `POST /instances/{{id}}/rules` \u2192 add interception rules\n3. `GET  /instances/{{id}}/cert` \u2192 download CA cert\n4. Point your client at `host:{{port}}` as its HTTP proxy\n\n**OpenAPI contract**: [openapi.json](/openapi.json)\n",
        "version": "2.0.0"
    },
    "paths": {
...
        "/instances": {
            "get": {
                "tags": [
                    "Instances"
                ],
                "summary": "List Instances",
                "description": "List every active mitmproxy instance with summary info.",
                "operationId": "list_instances_instances_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InstanceSummary"
                                    },
                                    "title": "Response List Instances Instances Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Instances"
                ],
                "summary": "Create Instance",
                "description": "Spin up a new mitmproxy instance on the next available port.",
                "operationId": "create_instance_instances_post",
                "parameters": [
                    {
                        "name": "ttl",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Instance lifespan in seconds.  Defaults to INSTANCE_TTL env var.",
                            "title": "Ttl"
                        },
                        "description": "Instance lifespan in seconds.  Defaults to INSTANCE_TTL env var."
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateInstanceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/instances/{instance_id}": {
            "get": {
                "tags": [
                    "Instances"
                ],
                "summary": "Get Instance",
                "description": "Detailed information about a specific instance including its rules.",
                "operationId": "get_instance_instances__instance_id__get",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Instance Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstanceDetail"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Instances"
                ],
                "summary": "Delete Instance",
                "description": "Terminate a running instance and free its port.",
                "operationId": "delete_instance_instances__instance_id__delete",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Instance Id"
                        }
                    },
                    {
                        "name": "cleanup",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "description": "Also remove rule files and CA directory",
                            "default": false,
                            "title": "Cleanup"
                        },
                        "description": "Also remove rule files and CA directory"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/instances/{instance_id}/renew": {
            "post": {
                "tags": [
                    "Instances"
                ],
                "summary": "Renew Instance",
                "description": "Extend an instance's lifespan.  Resets the expiry clock.",
                "operationId": "renew_instance_instances__instance_id__renew_post",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Instance Id"
                        }
                    },
                    {
                        "name": "ttl",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "New TTL in seconds.  If omitted the instance's current TTL is reused.",
                            "title": "Ttl"
                        },
                        "description": "New TTL in seconds.  If omitted the instance's current TTL is reused."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RenewResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/instances/{instance_id}/rules": {
            "get": {
                "tags": [
                    "Rules"
                ],
                "summary": "List Rules",
                "description": "List all interception rules for an instance.",
                "operationId": "list_rules_instances__instance_id__rules_get",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Instance Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RuleResponse"
                                    },
                                    "title": "Response List Rules Instances  Instance Id  Rules Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rules"
                ],
                "summary": "Create Rule",
                "description": "Add a new interception rule.  Takes effect immediately.",
                "operationId": "create_rule_instances__instance_id__rules_post",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Instance Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RuleCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/instances/{instance_id}/rules/{rule_index}": {
            "delete": {
                "tags": [
                    "Rules"
                ],
                "summary": "Remove Rule",
                "description": "Remove a rule by its positional index.",
                "operationId": "remove_rule_instances__instance_id__rules__rule_index__delete",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Instance Id"
                        }
                    },
                    {
                        "name": "rule_index",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Rule Index"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/instances/{instance_id}/rules/{rule_index}/toggle": {
            "patch": {
                "tags": [
                    "Rules"
                ],
                "summary": "Toggle Rule Endpoint",
                "description": "Toggle a rule between enabled and disabled.",
                "operationId": "toggle_rule_endpoint_instances__instance_id__rules__rule_index__toggle_patch",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Instance Id"
                        }
                    },
                    {
                        "name": "rule_index",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Rule Index"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/instances/{instance_id}/cert": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Get Cert",
                "description": "Download the CA certificate PEM for client configuration.",
                "operationId": "get_cert_instances__instance_id__cert_get",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Instance Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateInstanceResponse": {
                "properties": {
                    "instanceId": {
                        "type": "string",
                        "title": "Instanceid"
                    },
                    "port": {
                        "type": "integer",
                        "title": "Port"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "ttl": {
                        "type": "integer",
                        "title": "Ttl"
                    },
                    "expiresAt": {
                        "type": "string",
                        "title": "Expiresat"
                    }
                },
                "type": "object",
                "required": [
                    "instanceId",
                    "port",
                    "status",
                    "ttl",
                    "expiresAt"
                ],
                "title": "CreateInstanceResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "InstanceSummary": {
                "properties": {
                    "instanceId": {
                        "type": "string",
                        "title": "Instanceid"
                    },
                    "port": {
                        "type": "integer",
                        "title": "Port"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "createdAt": {
                        "type": "string",
                        "title": "Createdat"
                    },
                    "uptimeSeconds": {
                        "type": "number",
                        "title": "Uptimeseconds"
                    },
                    "ttl": {
                        "type": "integer",
                        "title": "Ttl",
                        "description": "Configured lifespan in seconds"
                    },
                    "remainingSeconds": {
                        "type": "number",
                        "title": "Remainingseconds",
                        "description": "Seconds until auto-destruction"
                    },
                    "ruleCount": {
                        "type": "integer",
                        "title": "Rulecount"
                    },
                    "clientIps": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Clientips",
                        "description": "Unique client IP addresses that have connected through this proxy"
                    }
                },
                "type": "object",
                "required": [
                    "instanceId",
                    "port",
                    "status",
                    "createdAt",
                    "uptimeSeconds",
                    "ttl",
                    "remainingSeconds",
                    "ruleCount"
                ],
                "title": "InstanceSummary"
            },
            "MessageResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "MessageResponse"
            },
...
        }
    }
}